server/pyproject.toml
[tool.poetry]
name = "KonomiTV"
version = "0.1.0"
description = "KonomiTV: Kept Organized, Notably Optimized, Modern Interface TV media server"
package-mode = false

[tool.taskipy.tasks]
serve = "python KonomiTV.py"
dev = "python KonomiTV.py --reload"
lint = "ruff check --fix . ; pyright"
typecheck = "pyright"
update-thirdparty = "python -m misc.UpdateThirdparty latest"
build-and-serve = "cd ../client && yarn build && cd ../server && python KonomiTV.py"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
aerich = "^0.8.1"
aiofiles = "^24.1.0"
aiohttp = "^3.12.14"
ariblib = { url = "https://github.com/tsukumijima/ariblib/releases/download/v0.1.4/ariblib-0.1.4-py3-none-any.whl" }
bcrypt = "^4.3.0"
beautifulsoup4 = "^4.12.3"
biim = { url = "https://github.com/tsukumijima/biim/releases/download/v1.11.0-post1/biim-1.11.0.post1-py3-none-any.whl" }
colorama = "^0.4.6"
curl-cffi = "^0.13.0"
elevate = "^0.1.3"
fastapi = "^0.115.6"
fastapi-utils = "^0.8.0"
hashids = "^1.3.1"
httpx = {version = "^0.28.1", extras = ["http2"]}
opencv-python-headless = "^4.11.0.86"
passlib = {version = "^1.7.4", extras = ["bcrypt"]}
pillow = "^11.3.0"
ping3 = "^4.0.8"
psutil = "^5.9.6"
puremagic = "^1.28"
py7zr = "^0.22.0"
pydantic = "^2.10.3"
pymediainfo = "^6.1.0"
pyright = "^1.1.403"
python-jose = {version = "^3.4.0", extras = ["cryptography"]}
python-multipart = "^0.0.20"
pywin32 = {version = "^311", platform = "win32"}
requests = "^2.32.4"
rich = "^13.9.4"
"ruamel.yaml" = "^0.18.10"
ruff = "^0.9.1"
sse-starlette = "^2.1.3"
taskipy = "^1.14.1"
tortoise-orm = "^0.23.0"
typer = "^0.15.4"
typing-extensions = "^4.12.2"
typing-inspect = "^0.9.0"
tweepy = "^4.14.0"
tweepy-authlib = "^1.6.0"
tzdata = "^2024.2"
uvicorn = {version = "^0.34.0", extras = ["standard"]}
uvloop = {version = "^0.21.0", platform = "linux"}
watchfiles = "^1.0.4"
winloop = {version = "^0.1.8", platform = "win32"}
discord-py = "^2.5.2"

[tool.poetry.group.dev.dependencies]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.aerich]
tortoise_orm = "app.constants.DATABASE_CONFIG"
location = "./app/migrations/"
src_folder = "./"

[tool.ruff]
# インデントの幅を4スペースに設定
indent-width = 4
# Python 3.11 を利用する
target-version = "py311"
# data/, thirdparty/ ディレクトリは対象から除外する
extend-exclude = ["data/", "thirdparty/"]

[tool.ruff.lint]
# flake8, pycodestyle, pyupgrade, isort, Ruff 固有のルールを使う
select = ["F", "E", "W", "UP", "I", "RUF", "TID251"]
ignore = [
    "E501",    # 1行の長さを超えている場合の警告を抑制
    "E731",    # Do not assign a `lambda` expression, use a `def` を抑制
    "RUF001",  # 全角記号など `ambiguous unicode character` も使いたいため
    "RUF002",  # 全角記号など `ambiguous unicode character` も使いたいため
    "RUF003",  # 全角記号など `ambiguous unicode character` も使いたいため
    "RUF012",
]

[tool.ruff.lint.isort]
# インポートブロックの後に2行空ける
lines-after-imports = 2

[tool.ruff.lint.flake8-tidy-imports.banned-api]
# Python 3.11 + Pydantic で TypedDict を扱う際は、typing_extensions.TypedDict を使う必要がある
# ref: https://github.com/langchain-ai/langgraph/pull/2910
"typing.TypedDict".msg = "Use typing_extensions.TypedDict instead."

[tool.pyright]
# Python バージョンを指定
pythonVersion = "3.11"
# TypeCheckingMode を strict に設定（VSCode の設定と同等）
typeCheckingMode = "strict"
# プロジェクトルートを指定
pythonPlatform = "All"
# インクルードするパッケージパス
include = ["app/"]
# 除外するパス
exclude = [
    "**/__pycache__",
    "**/.venv",
    "**/node_modules",
    "data/",
    "logs/",
]
# VSCode の settings.json の diagnosticSeverityOverrides に対応する設定
reportConstantRedefinition = "none"
reportDeprecated = "warning"
reportMissingTypeStubs = "none"
reportPrivateImportUsage = "none"
reportShadowedImports = "none"
reportUnnecessaryComparison = "none"
reportUnknownArgumentType = "none"
reportUnknownMemberType = "none"
reportUnknownVariableType = "none"
reportUnusedFunction = "none"
